rename kinesisConsumer to messageConsumer
