+++ example/spec/parallel/kinesisTests/KinesisTestTriggerSpec.js
@@ -123,9 +123,9 @@ const ruleOverride = {
 const s3data = ['@cumulus/test-data/granules/L2_HR_PIXC_product_0001-of-4154.h5'];
 
 
-// When kinesis-type rules exist, the Cumulus lambda kinesisConsumer is
+// When kinesis-type rules exist, the Cumulus lambda messageConsumer is
 // configured to trigger workflows when new records arrive on a Kinesis
-// stream. When a record appears on the stream, the kinesisConsumer lambda
+// stream. When a record appears on the stream, the messageConsumer lambda
 // triggers workflows associated with the kinesis-type rules.
 describe('The Cloud Notification Mechanism Kinesis workflow', () => {
   const maxWaitForSFExistSecs = 60 * 4;
