+++ packages/deployment/app/cloudformation.template.yml
@@ -70,9 +70,9 @@ Resources:
           Fn::GetAtt:
             - dbIndexerLambdaFunction
             - Arn
-        kinesisConsumerLambdaFunction:
+        messageConsumerLambdaFunction:
           Fn::GetAtt:
-            - kinesisConsumerLambdaFunction
+            - messageConsumerLambdaFunction
             - Arn
         CreateReconciliationReportLambdaFunction:
           Fn::GetAtt:
