+++ packages/deployment/app/cumulus_api.template.yml
@@ -29,9 +29,9 @@ Parameters:
   dbIndexerLambdaFunction:
     Type: String
     Description: 'Arn for dbIndexer lambda function'
-  kinesisConsumerLambdaFunction:
+  messageConsumerLambdaFunction:
     Type: String
-    Description: 'kinesisConsumer lambda function arn'
+    Description: 'messageConsumer lambda function arn'
   CreateReconciliationReportLambdaFunction:
     Type: String
     Description: 'CreateReconciliationReport lambda function arn'
