+++ packages/integration-tests/index.js
@@ -291,7 +291,7 @@ function setProcessEnvironment(stackName, bucketName) {
   process.env.internal = bucketName;
   process.env.bucket = bucketName;
   process.env.stackName = stackName;
-  process.env.kinesisConsumer = `${stackName}-kinesisConsumer`;
+  process.env.messageConsumer = `${stackName}-messageConsumer`;
   process.env.KinesisInboundEventLogger = `${stackName}-KinesisInboundEventLogger`;
   process.env.CollectionsTable = `${stackName}-CollectionsTable`;
   process.env.ProvidersTable = `${stackName}-ProvidersTable`;
